Renamed Commander.execute! to Commander.execute?
I think it's a bit clearer about what this method is
supposed to do. Most of the time, bang methods should
return nil in Ruby, and be a more dangerous version of
their non-bang equivalents. In this case, we were using
it essentially to just specify output format (boolean or
string), so I renamed it to the predicate style which
is the more Ruby way of declaring that. I think.